+++ plugins/permission-node/package.json
@@ -8,7 +8,6 @@
   "publishConfig": {
     "access": "public",
     "main": "dist/index.cjs.js",
-    "module": "dist/index.esm.js",
     "types": "dist/index.d.ts"
   },
   "homepage": "https://backstage.io",
@@ -32,7 +31,7 @@
   "dependencies": {
     "@backstage/plugin-auth-backend": "^0.4.6",
     "@backstage/plugin-permission-common": "^0.1.0",
-    "@types/express": "*",
+    "@types/express": "^4.17.6",
     "express": "^4.17.1",
     "zod": "^3.11.6"
   },
